functions/ml_scorer/requirements.txt
azure-functions
onnxruntime
numpy
